Anti_spoofing test
very close to finish, the sofware tells you need to blinck and move your head to side and it turns when you do do each one. it works fast. but problem is that it says unknown with detection but. and also every single time it recognises that person software just ends.